Refactor TUI model.go, handle SQLite PRAGMA errors, prevent DB double init, and add test coverage for core and tui
