+++ tui/settings.go
@@ -1,0 +1,349 @@
+package tui
+
+import (
+	"encoding/csv"
+	"encoding/json"
+	"fmt"
+	"focusd/storage"
+	"focusd/system"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strconv"
+	"strings"
+
+	tea "github.com/charmbracelet/bubbletea"
+)
+
+func (m Model) handleSettingsKey(key string) (Model, tea.Cmd) {
+	if m.settingsAddingBrowser {
+		switch key {
+		case "esc":
+			m.settingsAddingBrowser = false
+			m.settingsBrowserInput = ""
+		case "backspace":
+			if len(m.settingsBrowserInput) > 0 {
+				m.settingsBrowserInput = m.settingsBrowserInput[:len(m.settingsBrowserInput)-1]
+			}
+		case "enter":
+			if strings.TrimSpace(m.settingsBrowserInput) == "" {
+				m.addToast("Browser name is required", toastError)
+				return m, nil
+			}
+			if err := storage.AddCustomBrowser(m.settingsBrowserInput); err != nil {
+				m.addToast(err.Error(), toastError)
+			} else {
+				m.addToast("Added " + m.settingsBrowserInput, toastSuccess)
+			}
+			m.settingsAddingBrowser = false
+			m.settingsBrowserInput = ""
+		default:
+			if len(key) == 1 && len(m.settingsBrowserInput) < 32 {
+				m.settingsBrowserInput += key
+			}
+		}
+		return m, nil
+	}
+
+	switch key {
+	case "j", "down":
+		m.settingsSelected = min(m.settingsSelected+1, 12)
+	case "k", "up":
+		m.settingsSelected = max(0, m.settingsSelected-1)
+	case "h", "left":
+		switch m.settingsSelected {
+		case 8:
+			m.settingsBrowserSelected = max(0, m.settingsBrowserSelected-1)
+		case 4:
+			current := storage.GetTrackingIntervalSeconds()
+			next := max(storage.MinTrackingIntervalSeconds, current-1)
+			if next != current {
+				storage.SetTrackingIntervalSeconds(next)
+				m.addToast(fmt.Sprintf("Tracking interval: %ds", next), toastSuccess)
+			}
+		case 5:
+			current := storage.GetIdleThresholdSeconds()
+			next := max(storage.MinIdleThresholdSeconds, current-5)
+			if next != current {
+				storage.SetIdleThresholdSeconds(next)
+				m.addToast(fmt.Sprintf("Idle threshold: %ds", next), toastSuccess)
+			}
+		case 7:
+			current := storage.GetWarningThresholdPercent()
+			next := max(storage.MinWarningThresholdPercent, current-5)
+			if next != current {
+				storage.SetWarningThresholdPercent(next)
+				m.addToast(fmt.Sprintf("Warning threshold: %d%%", next), toastSuccess)
+			}
+		}
+	case "l", "right":
+		switch m.settingsSelected {
+		case 8:
+			browsers := storage.GetCustomBrowsersList()
+			m.settingsBrowserSelected = min(m.settingsBrowserSelected+1, max(0, len(browsers)-1))
+		case 4:
+			current := storage.GetTrackingIntervalSeconds()
+			next := min(storage.MaxTrackingIntervalSeconds, current+1)
+			if next != current {
+				storage.SetTrackingIntervalSeconds(next)
+				m.addToast(fmt.Sprintf("Tracking interval: %ds", next), toastSuccess)
+			}
+		case 5:
+			current := storage.GetIdleThresholdSeconds()
+			next := min(storage.MaxIdleThresholdSeconds, current+5)
+			if next != current {
+				storage.SetIdleThresholdSeconds(next)
+				m.addToast(fmt.Sprintf("Idle threshold: %ds", next), toastSuccess)
+			}
+		case 7:
+			current := storage.GetWarningThresholdPercent()
+			next := min(storage.MaxWarningThresholdPercent, current+5)
+			if next != current {
+				storage.SetWarningThresholdPercent(next)
+				m.addToast(fmt.Sprintf("Warning threshold: %d%%", next), toastSuccess)
+			}
+		}
+	case "space", "enter":
+		switch m.settingsSelected {
+		case 0:
+			if m.daemonActive {
+				exec.Command("cmd", "/c", "taskkill", "/IM", system.DaemonProcessName, "/F").Start()
+				m.addToast("Daemon stop requested", toastWarning)
+			} else {
+				m.addToast("Run focusd start to launch daemon", toastInfo)
+			}
+		case 1:
+			enabled, _, _ := system.GetAutoStartEnabled()
+			if enabled {
+				system.DisableAutoStart()
+				m.addToast("Auto-start disabled", toastSuccess)
+			} else {
+				system.EnableAutoStart()
+				m.addToast("Auto-start enabled", toastSuccess)
+			}
+		case 2:
+			paused := storage.IsPaused()
+			storage.SetPaused(!paused)
+			m.addToast("Browser tracking toggled", toastSuccess)
+		case 3:
+			m.addToast("Smart grouping toggled", toastInfo)
+		case 4:
+			current := storage.GetTrackingIntervalSeconds()
+			next := min(storage.MaxTrackingIntervalSeconds, current+1)
+			if next != current {
+				storage.SetTrackingIntervalSeconds(next)
+			}
+			m.addToast(fmt.Sprintf("Tracking interval: %ds", next), toastSuccess)
+		case 5:
+			current := storage.GetIdleThresholdSeconds()
+			next := min(storage.MaxIdleThresholdSeconds, current+5)
+			if next != current {
+				storage.SetIdleThresholdSeconds(next)
+			}
+			m.addToast(fmt.Sprintf("Idle threshold: %ds", next), toastSuccess)
+		case 6:
+			enabled := system.GetBreakReminderEnabled()
+			minutes := system.GetBreakReminderMinutes()
+			if err := system.SetBreakReminder(!enabled, minutes); err != nil {
+				m.addToast("Failed to toggle focus alerts", toastError)
+			} else {
+				m.addToast("Focus alerts toggled", toastSuccess)
+			}
+		case 7:
+			current := storage.GetWarningThresholdPercent()
+			next := min(storage.MaxWarningThresholdPercent, current+5)
+			if next != current {
+				storage.SetWarningThresholdPercent(next)
+			}
+			m.addToast(fmt.Sprintf("Warning threshold: %d%%", next), toastSuccess)
+		case 8:
+			m.settingsAddingBrowser = true
+			m.settingsBrowserInput = ""
+		case 9:
+			openDBFolder()
+		case 10:
+			path, err := exportData(false)
+			if err != nil {
+				m.addToast("Export failed", toastError)
+			} else {
+				m.addToast("Exported to "+filepath.Base(path), toastSuccess)
+			}
+		case 11:
+			if err := launchGitHubUpdate(); err != nil {
+				m.addToast("Failed to launch updater", toastError)
+			} else {
+				m.addToast("Updater opened in new window", toastSuccess)
+			}
+		case 12:
+			m.modal = modal{
+				Active: true, Title: "WIPE DATA", Required: "DELETE", Danger: true,
+				Message: "This will delete all tracking data and cannot be undone.",
+			}
+		}
+	case "d":
+		if m.settingsSelected == 8 {
+			browsers := storage.GetCustomBrowsersList()
+			if len(browsers) > 0 {
+				i := min(m.settingsBrowserSelected, len(browsers)-1)
+				storage.RemoveCustomBrowser(browsers[i])
+				m.addToast("Removed "+browsers[i], toastWarning)
+			}
+		}
+	}
+	return m, nil
+}
+
+func (m Model) renderSettings(width, height int) string {
+	auto, _, _ := system.GetAutoStartEnabled()
+	paused := storage.IsPaused()
+	trackingInterval := storage.GetTrackingIntervalSeconds()
+	idleThreshold := storage.GetIdleThresholdSeconds()
+	focusAlerts := system.GetBreakReminderEnabled()
+	warningThreshold := storage.GetWarningThresholdPercent()
+	browsers := append([]string{"chrome.exe", "firefox.exe", "edge.exe"}, storage.GetCustomBrowsersList()...)
+	dbPath := m.dashboard.DBPath
+	if dbPath == "" {
+		dbPath, _ = storage.GetDBPath()
+	}
+	inner := max(20, width-6)
+	browserValue := truncate(strings.Join(browsers, "   "), max(12, inner-42)) + "   " + buttonText("+ Add Browser")
+	if m.settingsAddingBrowser {
+		browserValue = "New browser: [ " + cyanStyle.Render(padRight(m.settingsBrowserInput, 18)) + " ]"
+	}
+	lines := []string{
+		"DAEMON",
+		settingRow(0, m.settingsSelected, "Background service", statusPill(m.daemonActive)+"  "+buttonText(ifThen(m.daemonActive, "Stop", "Start")), inner),
+		settingRow(1, m.settingsSelected, "Auto-start on login", toggleText(auto), inner),
+		"",
+		"TRACKING",
+		settingRow(2, m.settingsSelected, "Browser tracking", toggleText(!paused), inner),
+		settingRow(3, m.settingsSelected, "Smart app grouping", toggleText(true)+"  "+mutedStyle.Render("(experimental)"), inner),
+		settingRow(4, m.settingsSelected, "Tracking interval", fmt.Sprintf("[ %d ] seconds", trackingInterval), inner),
+		settingRow(5, m.settingsSelected, "Idle threshold", fmt.Sprintf("[ %d ] seconds", idleThreshold), inner),
+		"",
+		"NOTIFICATIONS",
+		settingRow(6, m.settingsSelected, "Focus complete alert", toggleText(focusAlerts), inner),
+		settingRow(7, m.settingsSelected, "Limit warning at", fmt.Sprintf("[ %d ] %% usage", warningThreshold), inner),
+		"",
+		"BROWSERS",
+		settingRow(8, m.settingsSelected, "Tracked browsers", browserValue, inner),
+		"",
+		"DATA",
+		settingRow(9, m.settingsSelected, "Database path", mutedStyle.Render(truncate(dbPath, max(10, inner-45)))+"   "+buttonText("Open Folder"), inner),
+		settingRow(10, m.settingsSelected, "Export data", buttonText("Export CSV")+"   "+buttonText("Export JSON"), inner),
+		settingRow(11, m.settingsSelected, "Update app", buttonText("Update from GitHub"), inner),
+		settingRow(12, m.settingsSelected, "Danger zone", redStyle.Bold(true).Render("[ Uninstall / Wipe All Data ]"), inner),
+	}
+	for i, line := range lines {
+		switch line {
+		case "DAEMON", "TRACKING", "NOTIFICATIONS", "BROWSERS", "DATA":
+			lines[i] = boldStyle.Render(line) + "\n" + mutedStyle.Render("────────────────────────────────────────")
+		}
+	}
+	bodyLines := strings.Split(strings.Join(lines, "\n"), "\n")
+	maxBodyLines := max(6, height-6)
+	if len(bodyLines) > maxBodyLines {
+		selectedLine := 0
+		for i, line := range bodyLines {
+			if strings.Contains(line, ">") {
+				selectedLine = i
+				break
+			}
+		}
+		start := scrollStart(selectedLine, maxBodyLines-2, len(bodyLines))
+		end := min(len(bodyLines), start+maxBodyLines-2)
+		window := bodyLines[start:end]
+		if start > 0 {
+			window = append([]string{mutedStyle.Render("↑ more")}, window...)
+		}
+		if end < len(bodyLines) {
+			window = append(window, mutedStyle.Render("↓ more"))
+		}
+		bodyLines = window
+	}
+	return panel("SETTINGS", "", width-2, "\n"+strings.Join(bodyLines, "\n")+"\n", true)
+}
+
+func settingRow(index, selected int, label, value string, width int) string {
+	cursor := "  "
+	if index == selected {
+		cursor = cyanStyle.Render("> ")
+	}
+	labelWidth := min(24, max(12, width/3))
+	line := cursor + padRight(label, labelWidth) + value
+	line = fitLine(line, width)
+	if index == selected {
+		return selectedRowStyle.Render(padRight(line, width))
+	}
+	return line
+}
+
+func statusPill(running bool) string {
+	if running {
+		return "[ " + greenStyle.Render("● RUNNING") + " ]"
+	}
+	return "[ " + redStyle.Render("○ STOPPED") + " ]"
+}
+
+func toggleText(on bool) string {
+	if on {
+		return "[" + greenStyle.Render("✓") + "] Enabled"
+	}
+	return "[" + mutedStyle.Render("·") + "] Disabled"
+}
+
+func buttonText(label string) string {
+	return cyanStyle.Render("[ " + label + " ]")
+}
+
+func openDBFolder() {
+	if dbPath, err := storage.GetDBPath(); err == nil {
+		exec.Command("explorer.exe", filepath.Dir(dbPath)).Start()
+	}
+}
+
+func launchGitHubUpdate() error {
+	exePath, err := os.Executable()
+	if err != nil {
+		return err
+	}
+	cmd := exec.Command("cmd", "/c", "start", "", exePath, "update")
+	return cmd.Start()
+}
+
+func exportData(jsonOut bool) (string, error) {
+	userProfile := os.Getenv("USERPROFILE")
+	exportDir := filepath.Join(userProfile, "Desktop")
+	if userProfile == "" {
+		exportDir = "."
+	}
+	if jsonOut {
+		path := filepath.Join(exportDir, "focusd_export.json")
+		apps, err := storage.GetAllAppStats()
+		if err != nil {
+			return "", err
+		}
+		b, err := json.MarshalIndent(apps, "", "  ")
+		if err != nil {
+			return "", err
+		}
+		return path, os.WriteFile(path, b, 0644)
+	}
+	path := filepath.Join(exportDir, "focusd_export.csv")
+	apps, err := storage.GetAllAppStats()
+	if err != nil {
+		return "", err
+	}
+	f, err := os.Create(path)
+	if err != nil {
+		return "", err
+	}
+	defer f.Close()
+	w := csv.NewWriter(f)
+	defer w.Flush()
+	w.Write([]string{"Date", "App Name", "Executable", "Duration (seconds)", "Open Count"})
+	for _, app := range apps {
+		w.Write([]string{app.Date, app.AppName, app.ExeName, strconv.Itoa(app.TotalDurationSecs), strconv.Itoa(app.OpenCount)})
+	}
+	return path, nil
+}
